requirements.txt
boto3==1.20.15
ExifRead==2.3.2
Flask==2.0.2
Flask-Ext==0.1
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Passlib==0.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.1.3
Flask-WTF==0.15.1
Jinja2==3.0.2
jsonschema==4.2.1
Pillow==8.4.0
psycopg2==2.9.1
pyrsistent==0.18.0
SQLAlchemy==1.4.26
urllib3==1.26.7
WTForms==2.3.3